requirements.txt
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.7
django-appconf==1.0.5
django-googledrive-storage==1.5.2
django-heroku==0.3.1
django-ipware==4.0.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
oauth2client==4.1.3
Pillow==8.3.2
protobuf==3.18.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
ratelim==0.1.6
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7
whitenoise==5.3.0
